- make height uint64  - remove redundant isValidated bool

+++ das/worker.go
@@ -135,7 +135,7 @@ func (w *worker) sample(ctx context.Context, timeout time.Duration, height uint6
 	if w.state.isRecentHeader {
 		err = w.broadcast(ctx, shrexsub.Notification{
 			DataHash: h.DataHash.Bytes(),
-			Height:   h.Height(),
+			Height:   uint64(h.Height()),
 		})
 		if err != nil {
 			log.Warn("failed to broadcast availability message",
